Add some real basic layer support
This is one of the first steps into making the game play the way it
should. I still need to figure out how I am going to add an animation
for layer transition.

+++ src/matrix/Vec2.java
@@ -118,7 +118,7 @@ public class Vec2 {
 		y = vec.y;
 	}
 
-	public Vec3 asVec3(int z) {
+	public Vec3 asVec3(float z) {
 
 		return new Vec3(x, y, z);
 	}
